Add haproxy_ssl_letsencrypt_enable support for RHEL
This commit adds the EPEL `certbot` package and its dependencies to
the `haproxy_distro_certbot_packages` variable in order to facilitate
automatic issuing of the Let's Encrypt certificates on RHEL-based
distros.

It also adds the steps necessary to enable the EPEL repository on the
HAProxy hosts because it is not enabled by default.

Upstream-Status: Pending
Signed-off-by: Stephanos Ioannidis <root@stephanos.io>